pyproject.toml
[tool.poetry]
name = "agencia"
version = "0.1.0"
description = ""
authors = ["Pedro Henrique de Sousa Alcântara <alcantarapedro69@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.5"
drf-spectacular = "^0.24.2"
black = "^22.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
